+++ frontend/src/app/_shared/shared.module.ts
@@ -15,7 +15,7 @@ import { InViewportModule } from 'ng-in-viewport';
 import { Ng2AutoCompleteModule } from 'ng2-auto-complete';
 import { ApiErrorComponent } from '../api-error/api-error.component';
 import { SidebarConfigService } from '../sidebar/sidebar-config.service';
-import { TreeRulesComponent } from '../trees/forests/tree-guidelines/tree-rules.component';
+import { TreeRulesComponent } from '../trees/forests/tree-guidelines/tree-rules/tree-rules.component';
 import { SpinnerModule } from 'angular2-spinner/dist';
 
 @NgModule({
