+++ frontend/src/app/application-forms/tree-application-form/tree-application-form.component.html
@@ -2,7 +2,7 @@
   <div class="usa-grid">
     <nav class="breadcrumb">
       <ol>
-        <li><a routerLink="/christmas-trees/forests">Christmas tree permit forests</a></li>
+        <li><a routerLink="/christmas-trees/forests">Christmas tree permits</a></li>
         <li><a [routerLink]="'/christmas-trees/forests/' + forest.forestAbbr + '/tree-guidelines'">{{ forest.forestName }}</a></li>
         <li>Buy a permit</li>
       </ol>
