+++ frontend/src/app/application-forms/tree-application-form/tree-permit-view/tree-permit-view.component.html
@@ -1,7 +1,7 @@
 <section id="application-view" class="usa-section usa-grid">
   <nav class="breadcrumb">
     <ol>
-      <li><a routerLink="/christmas-trees/forests">Christmas tree permit forests</a></li>
+      <li><a routerLink="/christmas-trees/forests">Christmas tree permits</a></li>
       <li><a [routerLink]="'/christmas-trees/forests/' + forest.forestAbbr + '/tree-guidelines'">{{ forest.forestName }}</a></li>
       <li><a [routerLink]="'/applications/christmas-trees/forests/' + forest.forestAbbr + '/new'">Buy a permit</a></li>
       <li>Permit order confirmation</li>
