+++ frontend/src/app/trees/forests/forest-finder/forest-finder.component.html
@@ -2,9 +2,13 @@
   <div class="usa-grid">
     <nav class="breadcrumb">
       <ol>
-        <li>Christmas tree permit forests</li>
+        <li>Christmas tree permits</li>
       </ol>
     </nav>
+    <h1>Buy a Christmas tree permit</h1>
+    <p>National Forests have their own rules, season dates and permit prices for cutting down Christmas trees. Not all forests offer a Christmas tree
+cutting season. For those that do, you can buy your Christmas tree permits in person at the forest ranger district offices. Some forests offer
+the option to purchase Christmas tree permits online.</p>
     <form id="skip-nav-target" tabindex="-1" name='forestPickerForm' class="usa-forms" (submit)="goToForest(selectedForest.forestAbbr)" autocomplete="off">
       <label for="forest-finder-input">Find a Forest:</label>
       <input id='forest-finder-input'
