+++ frontend/src/app/trees/forests/tree-guidelines/tree-cutting-dates/tree-cutting-dates.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnChanges } from '@angular/core';
-import { FilterPipe } from '../../../_pipes/filter.pipe';
-import { LineBreakFormatterPipe } from '../../../_pipes/line-break-formatter.pipe';
-import { TreeDistrictsUtilService } from './tree-districts-util.service';
+import { FilterPipe } from '../../../../_pipes/filter.pipe';
+import { LineBreakFormatterPipe } from '../../../../_pipes/line-break-formatter.pipe';
+import { TreeDistrictsUtilService } from '../tree-districts-util.service';
 
 @Component({
   selector: 'app-tree-cutting-dates',
@@ -13,7 +13,11 @@ export class TreeCuttingDatesComponent implements OnChanges {
   districtsWithHoursAndDates: any = [];
   districtsWithPermits: any = [];
 
-  constructor(private filter: FilterPipe, private lineBreakFormatter: LineBreakFormatterPipe, public districtUtil: TreeDistrictsUtilService) {}
+  constructor(
+    private filter: FilterPipe,
+    private lineBreakFormatter: LineBreakFormatterPipe,
+    public districtUtil: TreeDistrictsUtilService
+  ) {}
 
   populateDistricts() {
     const locations = this.forest.locations;
@@ -22,12 +26,12 @@ export class TreeCuttingDatesComponent implements OnChanges {
 
     for (const key of Object.keys(districts)) {
       const district = districts[key];
-      const areaLocations = district.locations.filter( location => location.type.startsWith('cutting-area'));
+      const areaLocations = district.locations.filter(location => location.type.startsWith('cutting-area'));
       if (areaLocations.length) {
         district.locations = areaLocations;
         this.districtsWithHoursAndDates.push(district);
       }
-      const permitLocations = district.locations.filter( location => location.type === 'district-permits');
+      const permitLocations = district.locations.filter(location => location.type === 'district-permits');
       if (permitLocations.length) {
         district.locations = permitLocations;
         this.districtsWithPermits.push(district);
