+++ frontend/src/app/trees/forests/tree-guidelines/tree-guidelines.component.html
@@ -2,7 +2,7 @@
   <div class="usa-grid" *ngIf="forest">
     <nav class="breadcrumb">
       <ol>
-        <li><a routerLink="/christmas-trees/forests">Christmas tree permit forests</a></li>
+        <li><a routerLink="/christmas-trees/forests">Christmas tree permits</a></li>
         <li>{{ forest.forestName }}</li>
       </ol>
     </nav>
