+++ frontend/src/app/trees/forests/tree-guidelines/tree-rules/tree-rules.component.spec.ts
@@ -1,10 +1,10 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { FilterPipe } from '../../../_pipes/filter.pipe';
-import { forest } from '../../_mocks/forest.mock';
+import { FilterPipe } from '../../../../_pipes/filter.pipe';
+import { forest } from '../../../_mocks/forest.mock';
 import { TreeRulesComponent } from './tree-rules.component';
-import { LineBreakFormatterPipe } from '../../../_pipes/line-break-formatter.pipe';
-import { SpacesToDashesPipe } from '../../../_pipes/spaces-to-dashes.pipe';
-import { UtilService } from '../../../_services/util.service';
+import { LineBreakFormatterPipe } from '../../../../_pipes/line-break-formatter.pipe';
+import { SpacesToDashesPipe } from '../../../../_pipes/spaces-to-dashes.pipe';
+import { UtilService } from '../../../../_services/util.service';
 
 describe('TreeRulesComponent', () => {
   let component: TreeRulesComponent;
